[BugFix] Fix be bugs found bysonar

+++ be/src/exec/pipeline/aggregate/aggregate_distinct_blocking_source_operator.cpp
@@ -35,7 +35,7 @@ StatusOr<vectorized::ChunkPtr> AggregateDistinctBlockingSourceOperator::pull_chu
 
     // For having
     RETURN_IF_ERROR(eval_conjuncts_and_in_filters(_aggregator->conjunct_ctxs(), chunk.get()));
-    _aggregator->update_num_rows_returned(-(old_size - chunk->num_rows()));
+    _aggregator->update_num_rows_returned(-(old_size - static_cast<int64_t>(chunk->num_rows())));
 
     DCHECK_CHUNK(chunk);
     return std::move(chunk);
